Adjust and enable `VolatilePtr` tests
